Struct fault_log::FaultEntry

pub struct FaultEntry {
    pub fault_type: FaultType,
    pub error_code: Option<u64>,
    pub core: Option<u8>,
    pub running_task: Option<String>,
    pub running_app_crate: Option<String>,
    pub address_accessed: Option<VirtualAddress>,
    pub instruction_pointer: Option<VirtualAddress>,
    pub crate_error_occured: Option<String>,
    pub replaced_crates: Vec<String>,
    pub action_taken: RecoveryAction,
}

A data structure to hold information about each fault.

Fields
fault_type: FaultType

Type of fault

error_code: Option<u64>

Error code returned with the exception

core: Option<u8>

The core error occured

running_task: Option<String>

Task runnning immediately before the Exception

running_app_crate: Option<String>

If available the application crate that spawned the task

address_accessed: Option<VirtualAddress>

For page faults the address the program attempted to access. None for other faults

instruction_pointer: Option<VirtualAddress>

Address at which exception occured

crate_error_occured: Option<String>

Crate the address at which exception occured located

replaced_crates: Vec<String>

List of crates reloaded from memory to recover from fault

action_taken: RecoveryAction

Recovery Action taken as a result of the fault

Implementations

impl FaultEntry

pub fn new(fault_type: FaultType) -> FaultEntry

Returns an empty FaultEntry with only fault_type field filled.