Crate fault_log

This crate logs all the faults occuring within Theseus. Maintains a list of exceptions and panics that has occured since booting up. This crate does not hold reference to any task or app.

Structs

FaultEntry

A data structure to hold information about each fault.

Enums

FaultType

The possible faults (panics and exceptions) encountered during operations.

RecoveryAction

The different types of recovery procedures used for the observed fault

Functions

clear_fault_log

Clears the log of faults so far occured in the system

from_exception_number

Utility function to get Fault type from exception number.

get_the_most_recent_match

Provides the most recent entry in the log for given crate Utility function for iterative crate replacement

log_exception

Add a new exception instance to the fault log. Generally it will have fault_type and instruction_pointer. If error_code is provided with exception it will be send to the function as Some(error_code). If the exception is a page fault the address attempted to accesss will also be send to the function as Some(address_accessed)

log_handled_fault

Add a FaultEntry to fault log.

log_panic_entry

Add a new panic instance to the fault log.

print_fault_log

Prints the fault log

remove_unhandled_exceptions

Removes the unhandled faults from the fault log and returns. Is useful when we update the recovery detail about unhandled exceptions.